requirements.txt
pytest~=5.4.3
pynput~=1.6.8
paramiko~=2.7.1
scp~=0.13.2
pandas~=1.0.5
psutil~=5.7.2
TestLink-API-Python-client~=0.8.1
distro~=1.5.0
pytest-html==2.1.1
ansi2html==1.5.2